Revert "Nytt oppsett for gauge metrikker"
This reverts commit 87bbcedf3bb3e711f8c96bfd0bd76530f9f9ac91.
